# src/domain/models/users/user.ts
import { ApiProperty } from '@nestjs/swagger';

export class UserModel {
  @ApiProperty()
  id?: number;
  @ApiProperty()
  name: string;
  @ApiProperty()
  email: string;
  @ApiProperty()
  password: string;
}